wsd: move special brokers to a new home
