+++ wsd/COOLWSD.cpp
@@ -17,6 +17,7 @@
 #include "CommandControl.hpp"
 #endif
 #if !MOBILEAPP
+#include <wsd/SpecialBrokers.hpp>
 #include <HostUtil.hpp>
 #endif // !MOBILEAPP
 
@@ -120,7 +121,7 @@ using Poco::Net::PartHandler;
 #include <Clipboard.hpp>
 #include <Crypto.hpp>
 #include <DelaySocket.hpp>
-#include "DocumentBroker.hpp"
+#include <wsd/DocumentBroker.hpp>
 #include <common/JsonUtil.hpp>
 #include <common/FileUtil.hpp>
 #include <common/JailUtil.hpp>
